Pushed everything down a folder into wma
