+++ mlr-core/src/main/java/gov/usgs/wma/model/Geometry.java
@@ -1,4 +1,4 @@
-package gov.usgs.model;
+package gov.usgs.wma.model;
 
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
@@ -9,19 +9,20 @@ import javax.xml.bind.annotation.XmlEnum;
 import javax.xml.bind.annotation.XmlEnumValue;
 
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "Feature", propOrder
-	= {"type", "geometry", "pointProperties"
+@XmlType(name = "Geometry", propOrder
+	= {"type"
 	})
 
-@XmlRootElement(name = "Feature")
-public class Feature {
+@XmlRootElement(name = "Geometry")
+//@ApiModel(description = "GeoJSon geometry")
+public class Geometry {
 
 	@XmlType(name = "TypeEnum")
 	@XmlEnum(String.class)
 	public enum TypeEnum {
 
-		@XmlEnumValue("Feature")
-		FEATURE(String.valueOf("Feature"));
+		@XmlEnumValue("Point")
+		POINT(String.valueOf("Point"));
 
 		private String value;
 
@@ -49,16 +50,11 @@ public class Feature {
 	}
 
 	@XmlElement(name = "type")
+//	@ApiModelProperty(required = true, value = "the geometry type")
 	private TypeEnum type = null;
 
-	@XmlElement(name = "geometry")
-	private Point geometry = null;
-
-	@XmlElement(name = "pointProperties")
-	private PointProperties pointProperties = null;
-
 	/**
-	 * Get type
+	 * the geometry type
 	 *
 	 * @return type
   *
@@ -71,42 +67,12 @@ public class Feature {
 		this.type = type;
 	}
 
-	/**
-	 * Get geometry
-	 *
-	 * @return geometry
-  *
-	 */
-	public Point getGeometry() {
-		return geometry;
-	}
-
-	public void setGeometry(Point geometry) {
-		this.geometry = geometry;
-	}
-
-	/**
-	 * Get pointProperties
-	 *
-	 * @return pointProperties
-  *
-	 */
-	public PointProperties getPointProperties() {
-		return pointProperties;
-	}
-
-	public void setPointProperties(PointProperties pointProperties) {
-		this.pointProperties = pointProperties;
-	}
-
 	@Override
 	public String toString() {
 		StringBuilder sb = new StringBuilder();
-		sb.append("class Feature {\n");
+		sb.append("class Geometry {\n");
 
 		sb.append("    type: ").append(toIndentedString(type)).append("\n");
-		sb.append("    geometry: ").append(toIndentedString(geometry)).append("\n");
-		sb.append("    pointProperties: ").append(toIndentedString(pointProperties)).append("\n");
 		sb.append("}");
 		return sb.toString();
 	}
