+++ mlr-core/src/main/java/gov/usgs/wma/model/MonitoringLocation.java
@@ -1,4 +1,4 @@
-package gov.usgs.model;
+package gov.usgs.wma.model;
 
 import java.util.ArrayList;
 import java.util.List;
